Move write_csv method to csv_helpers
